+++ src/demo/SideNav.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-  import List from "$lib/components/List.svelte";
+  import Button from "$lib/components/Button.svelte";
+import List from "$lib/components/List.svelte";
+  import { GamepadButtons } from "$lib/index.js";
 
   const listItems = [
     'Main',
@@ -8,6 +10,12 @@
     'Settings',
     'Fifth',
   ];
+
+  let lstElement: List;
+
+  export function getListElement(): HTMLElement {
+    return lstElement.getElement();
+  }
 
   interface Props {
     open: boolean
@@ -27,13 +35,29 @@
   }
 </script>
 
-<div id="sidenav" class="sidenav" style:width={open ? '250px' : '0'}>
-  <button class="closebtn" onclick={() => {open = false;}}>&times;</button>
+<div
+    id="sidenav"
+    class="sidenav"
+    style:width={open ? '250px' : '0'}
+    >
+  <Button
+    cssclassWrapper="closebtn"
+    onpressed={() => {open = false;}}
+      inputMapping={{
+      name: 'close side nav',
+      gamepad: -1,
+      buttons: [GamepadButtons.LEFT],
+      keys: ['x'],
+    }}
+    >
+    &times;
+  </Button>
   <List
+    bind:this={lstElement}
     disabled={!open}
     onpressed={listItemPressed}
     bind:selectedIndex
     items={listItems}
-    wrap={false}>
+    wrap={true}>
   </List>
 </div>